requirements.txt
fastapi==0.117.1
httpx==0.28.1
langchain==0.3.27
langchain_core==0.3.76
langchain_mcp_adapters==0.1.10
langchain_ollama==0.3.8
langchain_openai==0.3.33
langgraph==0.6.7
matplotlib==3.10.6
mcp==1.14.1
numpy==2.3.3
opencv_python==4.11.0.86
pandas==2.3.2
Pillow==11.3.0
pydantic==2.11.9
PyYAML==6.0.2
typing_extensions==4.15.0
uvicorn==0.37.0
langgraph-checkpoint-sqlite>=2.0.0
grandalf==0.8
tqdm>=4.66
reportlab==4.4.4
mplcursors==0.7
nbformat==5.10.4
plotly==6.3.1
pyyaml==6.0.2
ipywidgets==8.1.7
anywidget==0.9.18